Modularizing code and also some refactoring
