+++ app/controllers/application_controller.rb
@@ -1,0 +1,30 @@
+# Copyright 2015 Google, Inc
+# Copyright 2019 Public Lab
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+require "mapknitterExporter"
+require "sinatra"
+require "sinatra/cors"
+
+set :allow_origin, "*"
+set :allow_methods, "GET,HEAD,POST"
+set :allow_headers, "content-type,if-modified-since"
+set :expose_headers, "location,link"
+
+require "json"
+require "yaml"
+require "erb"
+require "open-uri"
+require "fog/google.rb"
+require "fog/local.rb"
